Initial attempt at getting index partitions working.
Decided to replace the internal representation of the indices from using
tuples of labels and expected values to just using an OrderedDict.

Refactored some of the duplicate code in getindex and setindex! to minimize
the work needed for this feature.

Needs a lot more tests for each condition while still support both Dict and Tuple
queries.